Relacionando entidades

+++ src/main/java/com/portal/comercio/Services/ventasServices.java
@@ -7,9 +7,9 @@ import com.portal.comercio.Models.ventas;
 import com.portal.comercio.dto.responseDto;
 
 public interface ventasServices {
-    public responseDto getVentasId(int codigo);
+    public responseDto getVentasId(Long codigo);
     // public Page<ventas> getVentas(Pageable pageable);
-    public responseDto updateVentas(ventas ventas, int codigo);
+    public responseDto updateVentas(ventas ventas, Long codigo);
     public responseDto saveVentas(ventas ventas);
-    public responseDto updateEstado(ventas ventas, int codigo); 
+    public responseDto updateEstado(ventas ventas, Long codigo); 
 }
